requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.5
colorama==0.4.6
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
dj-config-url==0.1.1
dj-database-url==2.0.0
Django==4.1
django-bootstrap5==23.3
django-crispy-forms==1.14.0
django-debug-toolbar==4.1.0
django-taggit==4.0.0
django-taggit-labels==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
psycopg2==2.9.6
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3


